Moved firebase functions to specific folders to organize functions
