+++ app/firebase/order/createOrder.ts
@@ -1,4 +1,4 @@
-import firebase_app from "./config";
+import firebase_app from "../config";
 import {
   getFirestore,
   collection,
@@ -12,7 +12,7 @@ import {
 } from "firebase/firestore";
 
 import { NextResponse } from "next/server";
-import { Item } from "../types/types";
+import { Item } from "../../types/types";
 
 interface Params {
   sessionID: string;
